backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
pydantic>=2.0
eth-account>=0.13.0
PyJWT>=2.9.0
slowapi>=0.1.9
python-dotenv>=1.0.0
httpx>=0.28.0
